refactor: 패키지 정리, 필요하지 않은 코드 제거, aws 설정, .gitignore 오타 수정
- AWS S3 키값은 노출되면 치명적이기도 하고, 자동으로 락이걸릴 수 있어요. 때문에 제가 개인적으로 드릴게요.
