+++ src/main/resources/application.properties
@@ -12,10 +12,10 @@ spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 spring.jpa.show-sql=true
 
-aws.access-key=your_access_key
-aws.secret-key=your_secret_key
-aws.region=your_region
-aws.s3.bucket=your_bucket_name
+aws.s3.access-key=${ACCESS_KEY}
+aws.s3.secret-key=${SECRET_KEY}
+aws.s3.region=${REGION}
+aws.s3.bucket=${BUCKET}
 
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
